# Improve using `//` in XPath performance