@naitoh 2025-04-27
> I have some concerns about maintainability because cache invalidation bug is hard to debug. Namespace can be modified through add_attribute/delete_attribute and through `attr_reader :attributes`. It can bypass cache invalidation mechanism.
> 
> Here are some ideas of cache strategy.
> 
>     1. Use a temporary cache and pass it as a method parameter
> 
> 
> ```ruby
> def node_test
>   namespaces_cache = {}
>   # Cache scope is enclosed in this method.
>   # There is no need to worry about cache invalidation.
>   call node.namespaces(namespaces_cache) to use the cache
> end
> 
> def namespaces(cache = nil)
>   if cache
>     cache[self] ||= calculate_namespaces(cache)
>   else
>     calculate_namespaces
>   end
> end
> ```

For conservatism, I have changed my mind about adopting policy 1.
Thanks!