@kou 2025-04-28
I have a bit concern about adding a `cache` (positional) argument to only some public namespace related APIs.

I feel that it may a bit ad-hoc approach. I'm not object this approach but can we improve our API?

For example, how about keeping a cache per document?

```ruby
class XPathParser
  def parse path, nodeset
    path_stack = @parser.parse( path )
    nodeset.first.document.enable_cache do
      # New document level cache is created and available in this block.
      match( path_stack, nodeset )
    end
  end
end
```