@naitoh 2025-04-28
I add `Document#enable_namespaces_cache` method to control whether caching is enabled or disabled.
What do you think?