@naitoh 2025-04-30
I have corrected the points you pointed out.
What do you think?