@kou 2025-05-03
Let's revisit this after #252. This implementation will be changed by #252.