@naitoh 2025-05-08
I have corrected the points you pointed out.
What do you think?