@naitoh 2025-05-11
I have corrected the points you pointed out.
What do you think?